net: collapse most of server impls
This treats the basic server as a special case of the VLAN server where
there is only one VLAN, and no tag handling occurs.
