+++ task/net/src/server.rs
@@ -2,15 +2,26 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-use crate::{bsp_support, bsp_support::Bsp, idl};
+use crate::bsp_support;
+use crate::generated::{self, SOCKET_COUNT};
+use crate::{idl, ETH_IRQ, NEIGHBORS, WAKE_IRQ_BIT};
+
 use drv_stm32h7_eth as eth;
 use idol_runtime::RequestError;
-use smoltcp::wire::EthernetAddress;
 use task_net_api::{
     KszError, KszMacTableEntry, LargePayloadBehavior, MacAddress,
     ManagementCounters, ManagementLinkStatus, MgmtError, PhyError, RecvError,
     SendError, SocketName, UdpMetadata,
 };
+
+use core::iter::zip;
+use heapless::Vec;
+use smoltcp::iface::{Interface, Neighbor, SocketHandle, SocketStorage};
+use smoltcp::socket::UdpSocket;
+use smoltcp::wire::{
+    EthernetAddress, IpAddress, IpCidr, Ipv6Address, Ipv6Cidr,
+};
+use userlib::{sys_post, sys_refresh_task_id, UnwrapLite};
 
 /// Abstraction trait to reduce code duplication between VLAN and non-VLAN
 /// server implementations.
@@ -32,15 +43,15 @@ pub trait NetServer {
         metadata: UdpMetadata,
         payload: idol_runtime::Leased<idol_runtime::R, [u8]>,
     ) -> Result<(), RequestError<SendError>>;
-
-    fn eth_bsp(&mut self) -> (&eth::Ethernet, &mut Self::Bsp);
-
-    /// Returns the MAC address for port 0
-    fn base_mac_address(&self) -> &EthernetAddress;
 }
 
 /// Implementation of the Net Idol interface.
-impl<T: NetServer> idl::InOrderNetImpl for T {
+impl<B, E, const N: usize> idl::InOrderNetImpl for GenServerImpl<'_, B, E, N>
+where
+    Self: NetServer,
+    B: bsp_support::Bsp,
+    E: DeviceExt,
+{
     fn recv_packet(
         &mut self,
         msg: &userlib::RecvMessage,
@@ -240,5 +251,224 @@ impl<T: NetServer> idl::InOrderNetImpl for T {
         let (eth, bsp) = self.eth_bsp();
         let out = bsp.management_counters(eth).map_err(MgmtError::from)?;
         Ok(out)
+    }
+}
+
+pub trait DeviceExt: for<'d> smoltcp::phy::Device<'d> {
+    fn read_and_clear_activity_flag(&self) -> bool;
+}
+
+/// State for the running network server
+pub struct GenServerImpl<'a, B, E, const N: usize>
+where
+    E: DeviceExt,
+{
+    eth: &'a eth::Ethernet,
+
+    vlan_state: [VLanState<E>; N],
+    client_waiting_to_send: [bool; SOCKET_COUNT],
+    bsp: B,
+
+    mac: EthernetAddress,
+}
+
+struct VLanState<E>
+where
+    E: DeviceExt,
+{
+    socket_handles: [SocketHandle; SOCKET_COUNT],
+    iface: Interface<'static, E>,
+}
+
+impl<'a, B, E, const N: usize> GenServerImpl<'a, B, E, N>
+where
+    B: bsp_support::Bsp,
+    E: DeviceExt,
+{
+    /// Size of buffer that must be allocated to use `dispatch`.
+    pub const INCOMING_SIZE: usize = idl::INCOMING_SIZE;
+
+    /// Builds a new `ServerImpl`, using the provided storage space.
+    pub(crate) fn new(
+        eth: &'a eth::Ethernet,
+        mut ipv6_addr: Ipv6Address,
+        mut mac: EthernetAddress,
+        bsp: B,
+        storage: &'static mut [Storage; N],
+        sockets: generated::Sockets<'static, N>,
+        mut mkdevice: impl FnMut(usize) -> E,
+    ) -> Self {
+        // Local storage; this will end up owned by the returned ServerImpl.
+        let mut vlan_state: Vec<VLanState<E>, N> = Vec::new();
+
+        let start_mac = mac;
+        // Each of these is replicated once per VID. Loop over them in lockstep.
+        for (i, (sockets, storage)) in zip(sockets.0, storage).enumerate() {
+            // Make some types explicit to try and make this clearer.
+            let sockets: [UdpSocket<'_>; SOCKET_COUNT] = sockets;
+
+            let neighbor_cache =
+                smoltcp::iface::NeighborCache::new(&mut storage.neighbors[..]);
+
+            let builder = smoltcp::iface::InterfaceBuilder::new(
+                mkdevice(i),
+                &mut storage.sockets[..],
+            );
+
+            storage.net = Ipv6Cidr::new(ipv6_addr, 64).into();
+            let mut iface = builder
+                .hardware_addr(mac.into())
+                .neighbor_cache(neighbor_cache)
+                .ip_addrs(core::slice::from_mut(&mut storage.net))
+                .finalize();
+
+            // Associate sockets with this interface.
+            let socket_handles = sockets.map(|s| iface.add_socket(s));
+            // Bind sockets to their ports.
+            for (&h, port) in zip(&socket_handles, generated::SOCKET_PORTS) {
+                iface
+                    .get_socket::<UdpSocket<'_>>(h)
+                    .bind((ipv6_addr, port))
+                    .unwrap_lite();
+            }
+
+            vlan_state
+                .push(VLanState {
+                    socket_handles,
+                    iface,
+                })
+                .unwrap_lite();
+
+            // Increment the MAC and IP addresses so that each VLAN has
+            // a unique address.
+            ipv6_addr.0[15] += 1;
+            mac.0[5] += 1;
+        }
+
+        Self {
+            eth,
+            client_waiting_to_send: [false; SOCKET_COUNT],
+            vlan_state: vlan_state.into_array().unwrap_lite(),
+            bsp,
+            mac: start_mac,
+        }
+    }
+
+    pub(crate) fn poll(&mut self, t: u64) -> smoltcp::Result<crate::Activity> {
+        let t = smoltcp::time::Instant::from_millis(t as i64);
+        // Do not be tempted to use `Iterator::any` here, it short circuits and
+        // we really do want to poll all of them.
+        let mut ip = false;
+        let mut mac_rx = false;
+        for vlan in &mut self.vlan_state {
+            ip |= vlan.iface.poll(t)?;
+            // Test and clear our receive activity flag.
+            mac_rx |= vlan.iface.device().read_and_clear_activity_flag();
+        }
+
+        Ok(crate::Activity { ip, mac_rx })
+    }
+
+    /// Iterate over sockets, waking any that can do work.  A task can do work
+    /// if all of the (internal) VLAN sockets can receive a packet, since
+    /// we don't know which VLAN it will write to.
+    pub fn wake_sockets(&mut self) {
+        for i in 0..SOCKET_COUNT {
+            if (0..N).any(|v| {
+                let want_to_send = self.client_waiting_to_send[i];
+                let socket = self.get_socket_mut(i, v).unwrap();
+                socket.can_recv() || (want_to_send && socket.can_send())
+            }) {
+                let (task_id, notification) = generated::SOCKET_OWNERS[i];
+                let task_id = sys_refresh_task_id(task_id);
+                sys_post(task_id, notification);
+            }
+        }
+    }
+
+    pub fn wake(&self) {
+        self.bsp.wake(self.eth)
+    }
+
+    fn get_handle(
+        &self,
+        index: usize,
+        vlan_index: usize,
+    ) -> Option<SocketHandle> {
+        self.vlan_state
+            .get(vlan_index)?
+            .socket_handles
+            .get(index)
+            .cloned()
+    }
+
+    /// Gets the socket `index`. If `index` is out of range, returns
+    /// `None`. Panics if `vlan_index` is out of range, which should
+    /// never happen (because messages with invalid VIDs are dropped in
+    /// RxRing).
+    ///
+    /// Sockets are currently assumed to be UDP.
+    pub(crate) fn get_socket_mut(
+        &mut self,
+        index: usize,
+        vlan_index: usize,
+    ) -> Option<&mut UdpSocket<'static>> {
+        Some(
+            self.vlan_state[vlan_index]
+                .iface
+                .get_socket::<UdpSocket<'_>>(
+                    self.get_handle(index, vlan_index)?,
+                ),
+        )
+    }
+
+    fn eth_bsp(&mut self) -> (&eth::Ethernet, &mut B) {
+        (self.eth, &mut self.bsp)
+    }
+
+    fn base_mac_address(&self) -> &EthernetAddress {
+        &self.mac
+    }
+
+    pub(crate) fn set_client_waiting_to_send(&mut self, i: usize, f: bool) {
+        self.client_waiting_to_send[i] = f;
+    }
+}
+
+impl<B, E, const N: usize> idol_runtime::NotificationHandler
+    for GenServerImpl<'_, B, E, N>
+where
+    E: DeviceExt,
+{
+    fn current_notification_mask(&self) -> u32 {
+        // We're always listening for our interrupt or the wake (timer) irq
+        ETH_IRQ | 1 << WAKE_IRQ_BIT
+    }
+
+    fn handle_notification(&mut self, bits: u32) {
+        // Interrupt dispatch.
+        if bits & ETH_IRQ != 0 {
+            self.eth.on_interrupt();
+            userlib::sys_irq_control(ETH_IRQ, true);
+        }
+        // The wake IRQ is handled in the main `net` loop
+    }
+}
+
+type NeighborStorage = Option<(IpAddress, Neighbor)>;
+
+pub struct Storage {
+    neighbors: [NeighborStorage; NEIGHBORS],
+    sockets: [SocketStorage<'static>; SOCKET_COUNT],
+    net: IpCidr,
+}
+
+impl Default for Storage {
+    fn default() -> Self {
+        Self {
+            neighbors: Default::default(),
+            sockets: Default::default(),
+            net: Ipv6Cidr::default().into(),
+        }
     }
 }
