+++ task/net/src/server_basic.rs
@@ -8,164 +8,52 @@
 
 use drv_stm32h7_eth as eth;
 
+use crate::bsp_support;
+use crate::generated;
+use crate::server::{DeviceExt, GenServerImpl, NetServer, Storage};
 use core::cell::Cell;
-use idol_runtime::{ClientError, NotificationHandler, RequestError};
+use idol_runtime::{ClientError, RequestError};
 use mutable_statics::mutable_statics;
-use smoltcp::iface::{Interface, Neighbor, SocketHandle, SocketStorage};
-use smoltcp::socket::UdpSocket;
-use smoltcp::wire::{
-    EthernetAddress, IpAddress, IpCidr, Ipv6Address, Ipv6Cidr,
-};
+use smoltcp::wire::{EthernetAddress, Ipv6Address};
 use task_net_api::{
     LargePayloadBehavior, RecvError, SendError, SocketName, UdpMetadata,
 };
-use userlib::{sys_post, sys_refresh_task_id};
-
-use crate::generated::{self, SOCKET_COUNT};
-use crate::server::NetServer;
-use crate::{bsp_support, idl, ETH_IRQ, NEIGHBORS, WAKE_IRQ_BIT};
-
-type NeighborStorage = Option<(IpAddress, Neighbor)>;
 
 /// Grabs references to the server storage arrays.  Can only be called once!
-fn claim_server_storage_statics() -> (
-    &'static mut [NeighborStorage; NEIGHBORS],
-    &'static mut [SocketStorage<'static>; SOCKET_COUNT],
-    &'static mut [IpCidr; 1],
-) {
+fn claim_server_storage_statics() -> &'static mut [Storage; 1] {
     mutable_statics! {
-        static mut NEIGHBOR_CACHE_STORAGE: [NeighborStorage; NEIGHBORS] =
-            [|| Default::default(); _];
-        static mut SOCKET_STORAGE: [SocketStorage<'static>; SOCKET_COUNT] =
-            [|| Default::default(); _];
-        static mut IPV6_NET: [IpCidr; 1] = [|| Ipv6Cidr::default().into(); _];
+        static mut STORAGE: [Storage; 1] = [|| Default::default(); _];
     }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
 
-/// State for the running network server.
-pub struct ServerImpl<'a, B> {
-    socket_handles: [SocketHandle; SOCKET_COUNT],
-    client_waiting_to_send: [bool; SOCKET_COUNT],
-    iface: Interface<'static, Smol<'a>>,
-    bsp: B,
+pub type ServerImpl<'a, B> = GenServerImpl<'a, B, Smol<'a>, 1>;
+
+pub fn new<'a, B>(
+    eth: &'a eth::Ethernet,
+    ipv6_addr: Ipv6Address,
     mac: EthernetAddress,
+    bsp: B,
+) -> ServerImpl<'a, B>
+where
+    B: bsp_support::Bsp,
+{
+    ServerImpl::new(
+        eth,
+        ipv6_addr,
+        mac,
+        bsp,
+        claim_server_storage_statics(),
+        generated::construct_sockets(),
+        |_| Smol::from(eth),
+    )
 }
 
-impl<'a, B: bsp_support::Bsp> ServerImpl<'a, B> {
-    /// Size of buffer that must be allocated to use `dispatch`.
-    pub const INCOMING_SIZE: usize = idl::INCOMING_SIZE;
-
-    /// Builds a new `ServerImpl`, using the provided storage space.
-    pub fn new(
-        eth: &'a eth::Ethernet,
-        ipv6_addr: Ipv6Address,
-        mac: EthernetAddress,
-        bsp: B,
-    ) -> Self {
-        let eth = Smol::from(eth);
-        let (neighbor_cache_storage, socket_storage, ipv6_net) =
-            claim_server_storage_statics();
-        ipv6_net[0] = Ipv6Cidr::new(ipv6_addr, 64).into();
-        let neighbor_cache =
-            smoltcp::iface::NeighborCache::new(&mut neighbor_cache_storage[..]);
-        let mut iface =
-            smoltcp::iface::InterfaceBuilder::new(eth, &mut socket_storage[..])
-                .hardware_addr(mac.into())
-                .neighbor_cache(neighbor_cache)
-                .ip_addrs(&mut ipv6_net[..])
-                .finalize();
-
-        // Create sockets and associate them with the interface.
-        let sockets = generated::construct_sockets();
-        let mut socket_handles = [None; generated::SOCKET_COUNT];
-        for (socket, h) in sockets.0.into_iter().zip(&mut socket_handles) {
-            *h = Some(iface.add_socket(socket));
-        }
-        let socket_handles = socket_handles.map(|h| h.unwrap());
-        // Bind sockets to their ports.
-        for (&h, &port) in socket_handles.iter().zip(&generated::SOCKET_PORTS) {
-            iface
-                .get_socket::<UdpSocket>(h)
-                .bind((ipv6_addr, port))
-                .map_err(|_| ())
-                .unwrap();
-        }
-
-        Self {
-            socket_handles,
-            client_waiting_to_send: [false; SOCKET_COUNT],
-            iface,
-            bsp,
-            mac,
-        }
-    }
-
-    /// Calls `smoltcp`'s internal poll function on our interface
-    pub(crate) fn poll(&mut self, t: u64) -> smoltcp::Result<crate::Activity> {
-        let ip = self
-            .iface
-            .poll(smoltcp::time::Instant::from_millis(t as i64))?;
-
-        // Test and clear our receive activity flag.
-        let mac_rx = self.iface.device().mac_rx.take();
-
-        Ok(crate::Activity { ip, mac_rx })
-    }
-
-    /// Iterate over sockets, waking any that can do work.
-    pub fn wake_sockets(&mut self) {
-        // There's something to do! Iterate over sockets looking for work.
-        // TODO making every packet O(n) in the number of sockets is super
-        // lame; provide a Waker to fix this.
-        for i in 0..SOCKET_COUNT {
-            let want_to_send = self.client_waiting_to_send[i];
-            let socket = self.get_socket_mut(i).unwrap();
-            if socket.can_recv() || (want_to_send && socket.can_send()) {
-                // Make sure the owner knows about this. This can
-                // technically cause spurious wakeups if the owner is
-                // already waiting in our incoming queue to recv. Maybe we
-                // fix this later.
-                let (task_id, notification) = generated::SOCKET_OWNERS[i];
-                let task_id = sys_refresh_task_id(task_id);
-                sys_post(task_id, notification);
-            }
-        }
-    }
-
-    /// Gets the socket handle for socket `index`. If `index` is out of range,
-    /// returns `BadMessage`.
-    ///
-    /// You often want `get_socket_mut` instead of this, but since it claims
-    /// `self` mutably, it is sometimes useful to inline it by calling this
-    /// followed by `eth.get_socket`.
-    fn get_handle(&self, index: usize) -> Result<SocketHandle, ClientError> {
-        self.socket_handles
-            .get(index)
-            .cloned()
-            .ok_or(ClientError::BadMessageContents)
-    }
-
-    /// Gets the socket `index`. If `index` is out of range, returns
-    /// `BadMessage`.
-    ///
-    /// Sockets are currently assumed to be UDP.
-    fn get_socket_mut(
-        &mut self,
-        index: usize,
-    ) -> Result<&mut UdpSocket<'static>, ClientError> {
-        Ok(self.iface.get_socket::<UdpSocket>(self.get_handle(index)?))
-    }
-
-    /// Calls the `wake` function on the BSP, which handles things like
-    /// periodic logging and monitoring of ports.
-    pub fn wake(&self) {
-        self.bsp.wake(&self.iface.device().eth);
-    }
-}
-
-impl<B: bsp_support::Bsp> NetServer for ServerImpl<'_, B> {
+impl<B: bsp_support::Bsp> NetServer for GenServerImpl<'_, B, Smol<'_>, 1>
+where
+    B: bsp_support::Bsp,
+{
     type Bsp = B;
 
     /// Requests that a packet waiting in the rx queue of `socket` be delivered
@@ -191,8 +79,8 @@ impl<B: bsp_support::Bsp> NetServer for ServerImpl<'_, B> {
         }
 
         let socket = self
-            .get_socket_mut(socket_index)
-            .map_err(RequestError::Fail)?;
+            .get_socket_mut(0, socket_index)
+            .ok_or(RequestError::Fail(ClientError::BadMessageContents))?;
         loop {
             match socket.recv() {
                 Ok((body, endp)) => {
@@ -242,18 +130,18 @@ impl<B: bsp_support::Bsp> NetServer for ServerImpl<'_, B> {
         }
 
         let socket = self
-            .get_socket_mut(socket_index)
-            .map_err(RequestError::Fail)?;
+            .get_socket_mut(0, socket_index)
+            .ok_or(RequestError::Fail(ClientError::BadMessageContents))?;
         match socket.send(payload.len(), metadata.into()) {
             Ok(buf) => {
                 payload
                     .read_range(0..payload.len(), buf)
                     .map_err(|_| RequestError::went_away())?;
-                self.client_waiting_to_send[socket_index] = false;
+                self.set_client_waiting_to_send(socket_index, false);
                 Ok(())
             }
             Err(smoltcp::Error::Exhausted) => {
-                self.client_waiting_to_send[socket_index] = true;
+                self.set_client_waiting_to_send(socket_index, true);
                 Err(SendError::QueueFull.into())
             }
             Err(_e) => {
@@ -261,30 +149,6 @@ impl<B: bsp_support::Bsp> NetServer for ServerImpl<'_, B> {
                 Err(SendError::Other.into())
             }
         }
-    }
-
-    fn eth_bsp(&mut self) -> (&eth::Ethernet, &mut Self::Bsp) {
-        (&self.iface.device().eth, &mut self.bsp)
-    }
-
-    fn base_mac_address(&self) -> &EthernetAddress {
-        &self.mac
-    }
-}
-
-impl<B> NotificationHandler for ServerImpl<'_, B> {
-    fn current_notification_mask(&self) -> u32 {
-        // We're always listening for our interrupt or the wake (timer) irq
-        ETH_IRQ | 1 << WAKE_IRQ_BIT
-    }
-
-    fn handle_notification(&mut self, bits: u32) {
-        // Interrupt dispatch.
-        if bits & ETH_IRQ != 0 {
-            self.iface.device().eth.on_interrupt();
-            userlib::sys_irq_control(ETH_IRQ, true);
-        }
-        // The wake IRQ is handled in the main `net` loop
     }
 }
 
@@ -294,7 +158,7 @@ impl<B> NotificationHandler for ServerImpl<'_, B> {
 // We gotta newtype the Ethernet driver since we're not in its crate. (This
 // implementation was once in its crate but that became a little gross.)
 
-struct Smol<'d> {
+pub struct Smol<'d> {
     eth: &'d eth::Ethernet,
     mac_rx: Cell<bool>,
 }
@@ -308,7 +172,7 @@ impl<'d> From<&'d eth::Ethernet> for Smol<'d> {
     }
 }
 
-struct OurRxToken<'d>(&'d Smol<'d>);
+pub struct OurRxToken<'d>(&'d Smol<'d>);
 impl<'d> smoltcp::phy::RxToken for OurRxToken<'d> {
     fn consume<R, F>(
         self,
@@ -322,7 +186,7 @@ impl<'d> smoltcp::phy::RxToken for OurRxToken<'d> {
     }
 }
 
-struct OurTxToken<'d>(&'d Smol<'d>);
+pub struct OurTxToken<'d>(&'d Smol<'d>);
 impl<'d> smoltcp::phy::TxToken for OurTxToken<'d> {
     fn consume<R, F>(
         self,
@@ -375,5 +239,11 @@ impl<'d> smoltcp::phy::Device<'d> for Smol<'_> {
 
     fn capabilities(&self) -> smoltcp::phy::DeviceCapabilities {
         crate::ethernet_capabilities(self.eth)
+    }
+}
+
+impl DeviceExt for Smol<'_> {
+    fn read_and_clear_activity_flag(&self) -> bool {
+        self.mac_rx.take()
     }
 }
